padronizando códigos

+++ Back/src/main/java/com/devnari/contrataai/persistencia/DisponibilidadeDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.devnari.contrataai.model.auxiliares.Disponibilidade;
 
-public interface DisponibilidadeDao extends JpaRepository<Disponibilidade, Integer> {
+public interface DisponibilidadeDao extends JpaRepository<Disponibilidade, Long> {
 
 }
