+++ Back/src/main/java/com/devnari/contrataai/persistencia/EnderecoDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.devnari.contrataai.model.auxiliares.Endereco;
 
-public interface EnderecoDao extends JpaRepository<Endereco, Integer> {
+public interface EnderecoDao extends JpaRepository<Endereco, Long> {
 
 }
